// web/lib/types.ts
// Additional types for Next app outside of those generated by graphql-codegen in '__generated__'
import { Drop, GearList, GearListItem } from '../__generated__/graphql'
export type GearListWithItems = GearList & {
	items?: Array<GearListItem>
	itemCount?: number
}

// FullDrop = drop with lists that contain GearListItems
export type FullDrop = Omit<Drop, 'lists'> & {
	lists?: Array<GearListWithItems>
}